move plugin to view
