+++ core/player/src/view/plugins/index.ts
@@ -2,4 +2,3 @@ export { default as TemplatePlugin } from "./template-plugin";
 export { default as StringResolverPlugin } from "./string-resolver";
 export { default as ApplicabilityPlugin } from "./applicability";
 export { default as SwitchPlugin } from "./switch";
-export type { ViewPlugin } from "./plugin";
